Use STATIC_URL instead of MEDIA_URL
Because of the historical setup, STATIC_URL is used for nearly
everything, and although new Django sort of forces a distinction between
STATIC_URL and MEDIA_URL, we ignore django in this instance.

This commit makes the site use STATIC_URL for the majority of things,
and a specific MEDIA_URL (which is STATIC_URL + "/media/") for user
uploaded stuff.

This commit also removes the /media hack from urls.py, and moves the
media from /media to /static/media